flatten value module
